Added middle mouse scroll support on links (images)
I'm a big fan of using the scroll feature when clicking the middle mouse button. Previously when the results page had enough images to enable scroll, you had to find some space between them to middle click, otherwise if clicked on an image it would redirect to it's post page.